requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Face Detection & Alignment
mtcnn>=0.1.1

# Face Restoration & Enhancement
gfpgan>=1.3.8
basicsr>=1.4.2
facexlib>=0.3.0
realesrgan>=0.3.0

# Face Recognition - AdaFace
timm>=0.9.0
scipy>=1.11.0

# Vector Database
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Uncomment for GPU support

# Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.0

# Utilities
tqdm>=4.66.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
pandas>=2.1.0

# Image Processing
albumentations>=1.3.1
imageio>=2.31.0

# Logging & Monitoring
loguru>=0.7.2

# Encryption
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
